--- examples/carrot-script/Cargo.toml
@@ -1,15 +1,0 @@
-[workspace]
-resolver = "2"
-
-members = [
-  # Please don't remove the following line, we use it to automatically
-  # detect insertion point for newly generated crates.
-  # @@INSERTION_POINT@@
-  "contracts/carrot",
-  "tests",
-]
-
-[profile.release]
-overflow-checks = true
-strip = true
-codegen-units = 1
