--- examples/carrot-script/README.md
@@ -1,7 +1,0 @@
-# carrot-script
-
-TODO: Write this readme
-
-*This project was bootstrapped with [ckb-script-templates].*
-
-[ckb-script-templates]: https://github.com/cryptape/ckb-script-templates
