--- examples/carrot-script/contracts/carrot/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "carrot"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-ckb-std = "0.15.1"
