--- examples/carrot-script/contracts/carrot/src/error.rs
@@ -1,26 +1,0 @@
-use ckb_std::error::SysError;
-
-#[cfg(test)]
-extern crate alloc;
-
-#[repr(i8)]
-pub enum Error {
-    IndexOutOfBound = 1,
-    ItemMissing,
-    LengthNotEnough,
-    Encoding,
-    // Add customized errors here...
-    CarrotAttack, 
-}
-
-impl From<SysError> for Error {
-    fn from(err: SysError) -> Self {
-        match err {
-            SysError::IndexOutOfBound => Self::IndexOutOfBound,
-            SysError::ItemMissing => Self::ItemMissing,
-            SysError::LengthNotEnough(_) => Self::LengthNotEnough,
-            SysError::Encoding => Self::Encoding,
-            SysError::Unknown(err_code) => panic!("unexpected sys error {}", err_code),
-        }
-    }
-}
