--- examples/carrot-script/scripts/find_clang
@@ -1,31 +1,0 @@
-#!/usr/bin/env bash
-#
-# An utility script used to find a binary of clang 16+
-
-if [[ -n "${CLANG}" ]]; then
-  echo "${CLANG}"
-  exit 0
-fi
-
-CANDIDATES=("clang" "clang-16" "clang-17")
-
-BREW_PREFIX=$(brew --prefix 2> /dev/null)
-if [[ -n "${BREW_PREFIX}" ]]; then
-  CANDIDATES+=(
-    "${BREW_PREFIX}/opt/llvm/bin/clang"
-    "${BREW_PREFIX}/opt/llvm@16/bin/clang"
-    "${BREW_PREFIX}/opt/llvm@17/bin/clang"
-  )
-fi
-
-for candidate in ${CANDIDATES[@]}; do
-  OUTPUT=$($candidate -dumpversion 2> /dev/null | cut -d'.' -f 1)
-
-  if [[ $((OUTPUT)) -ge 16 ]]; then
-    echo "$candidate"
-    exit 0
-  fi
-done
-
->&2 echo "Cannot find clang of version 16+!"
-exit 1
