--- examples/carrot-script/tests/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "tests"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-ckb-testtool = "0.10.2"
-serde_json = "1.0"
