+++ examples/ckb-rust-script/Cargo.toml
@@ -5,6 +5,7 @@ members = [
   # Please don't remove the following line, we use it to automatically
   # detect insertion point for newly generated crates.
   # @@INSERTION_POINT@@
+  "contracts/carrot",
   "contracts/group_exec",
   "contracts/crypto-service-hash",
   "contracts/ipc-child",
