+++ examples/ckb-rust-script/contracts/carrot/.gitignore
