+++ examples/ckb-rust-script/contracts/carrot/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "carrot"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+ckb-std = "0.17.0"
+
+[features]
+library = []
+native-simulator = ["library", "ckb-std/native-simulator"]
