+++ examples/ckb-rust-script/contracts/carrot/src/lib.rs
@@ -1,0 +1,9 @@
+#![cfg_attr(not(feature = "library"), no_std)]
+#![allow(special_module_name)]
+#![allow(unused_attributes)]
+#[cfg(feature = "library")]
+mod main;
+#[cfg(feature = "library")]
+pub use main::program_entry;
+
+extern crate alloc;
