+++ examples/ckb-rust-script/tests/src/lib.rs
@@ -23,6 +23,9 @@ mod tests_error;
 #[cfg(test)]
 mod test_type_id;
 
+#[cfg(test)]
+mod test_carrot;
+
 // The exact same Loader code from capsule's template, except that
 // now we use MODE as the environment variable
 const TEST_ENV_VAR: &str = "MODE";
